+++ resources/views/layouts/admin/app.blade.php
@@ -18,6 +18,7 @@
 
     {{-- <link href="{{ asset('css/app.css') }}" rel="stylesheet"> --}}
 
+    @stack('styles')
     @yield('styles')
 
 </head>