+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGE LOG
 
+# 2021-11-05
+- Add Product Enter (proses barang masuk)
 # 2021-11-04
 
 - Add Laravolt/Indonesia for Indonesian region data